Add files via upload
This is a caffe layer write in Python. I intended to rewrite the softmax and softmaxwithloss layer in caffe which were wrote in C++ as an exercise, but failed. After a few times trainning, all the data and diff will be NAN. I don't know why. Is there anyone who can help me？